+++ components/pm-subscription-handler/Changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Read NFS associated with MG by using MGName and subName(DCAEGEN2-2993)
 * Lazy loading error for nfs in read API (DCAEGEN2-3029)
 * Delete subscription API by Name(DCAEGEN2-2821)
+* Update Administrative status for measurement group(DCAEGEN2-2820)
 
 ## [1.3.2]
 ### Changed